// core/src/table/new-api/table.api.ts
export type TableProps<RowType = unknown, ColumnType extends BaseColumnType<RowType> = BaseColumnType<RowType>> =
	| BaseTableProps<RowType, ColumnType>
	| InfiniteScrollTableProps<RowType, ColumnType>;

export interface InfiniteScrollTableProps<
	RowType = unknown,
	ColumnType extends BaseColumnType<RowType> = BaseColumnType<RowType>
> extends Omit<BaseTableProps<RowType, ColumnType>, "data" | "virtualScrollOptions"> {
	/**
	 * Data to display.
	 */
	data?: (RowType | undefined)[];

	/**
	 * Use this props to enable infinite scroll feature. See {@link InfiniteScrollOptions} for details.
	 */
	infiniteScrollOptions: InfiniteScrollOptions;
}

export interface BaseTableProps<RowType = unknown, ColumnType extends BaseColumnType<RowType> = BaseColumnType<RowType>>
	extends TableTemplateProps.BaseProps, Ref<HTMLDivElement> {
	/**
	 * Define columns of table.
	 */
	columns: ColumnType[];

	/**
	 * Data to display. By default, if no dataKey or dataGetter is defined in {@link BaseColumnType}, then data will be parsed using index order from columns definition.
	 */
	data?: RowType[];

	/**
	 * Display table as cards.
	 * @default false
	 */
	cardView?: boolean;

	/**
	 * If true, all event handler are suppressed, as well as making table looks disabled.
	 */
	disabled?: boolean;

	/**
	 * Either a string or number pointing to the field in the {@link RowType} object, or a function returning the key of the row.
	 * See {@link RowKeyGetter}.
	 */
	rowKey?: string | number | RowKeyGetter<RowType>;

	/**
	 * Define sort order and event handler for column click.
	 *
	 * *Note:* If columns are resizable, the column in {@link SortOptions.sortState} should be updated after resizing to avoid an incorrect state of the sortable header cell.
	 */
	sortOptions?: SortOptions<ColumnType>;

	/**
	 * Enable or supply an option object to override default virtualized behavior, or use additional props that is supported by react-virtualized.
	 */
	virtualScrollOptions?: true | VirtualScrollOptions;

	/**
	 * A handler to scroll to a specific node (row).
	 */
	scrollToNode?(handler: TableScrollToNodeHandler): void;

	/**
	 * Event handlers for DnD behavior.
	 */
	dragDropOptions?: TableDragDropOptions<RowType>;

	/**
	 * Event handlers for resizing columns behavior.
	 */
	columnResizingOptions?: ColumnResizingOptions<ColumnType>;

	/**
	 * Use this prop to override rendering of table. See {@link TableComponentRenderers} for details.
	 */
	componentRenderers?: Partial<TableComponentRenderers<RowType, ColumnType>>;

	/**
	 * Used to specify whether aria-attributes (such as aria-label) will be added to the table footer.
	 */
	hasFootContent?: boolean;

	/**
	 * Define event handlers for rows. It receives a RowType object as argument, and return a list of handler functions.
	 * See {@link RowEventHandlerGetter}.
	 *
	 * *Note:* The event handlers returned by rowEventHandlers can be overridden by the same event handlers in {@link TableRenderPropsType.BodyRowProps}.
	 */
	rowEventHandlers?: RowEventHandlerGetter<RowType>;

	/**
	 * Define styling for row. It's a function receiving RowType object and its index as argument, and return a {@link RowStyles} object.
	 *
	 * *Note:* The properties returned by rowStyling can be overridden by the same properties in {@link TableRenderPropsType.BodyRowProps}.
	 */
	rowStyling?: RowStyleGetter<RowType>;

	/**
	 * Similar to {@link rowStyling}, but return style of cell as key, and return an {@link CellStyles} object.
	 *
	 * *Note:* The properties returned by cellStyling can be overridden by the same properties in {@link TableRenderPropsType.BodyCellProps}.
	 */
	cellStyling?: CellStyleGetter<RowType, ColumnType>;

	/**
	 * Whether the body cell and corresponding header cell(s) will be highlighted when mousing over on body cell.
	 */
	cellHighlighting?: boolean;

	/**
	 * If true, the arrow navigation keyboard handlers are suppressed.
	 */
	disableArrowNavigation?: boolean;

	/**
	 * This is useful for validation of Table's data when leaving the table.
	 * @param event – the FocusEvent, useful to check if the newly focused element (e.g. relatedTarget) is inside Table or not.
	 */
	onBlur?(event: FocusEvent<HTMLDivElement>): void;
}

/**
 * Options for DnD handling in table
 */
export interface TableDragDropOptions<
	RowType = unknown,
	DragItem = TableRenderPropsType.DragObject<RowType>,
	DropResult = TableRenderPropsType.DropResult<RowType>,
	HoveredItem = TableRenderPropsType.HoveredObject<RowType>
> {
	/**
	 * Whether the row can be dragged or not.
	 */
	canDrag?(params: { dragItem: DragItem }): boolean;

	/**
	 * Handler when the row is being dropped.
	 */
	onDrop?(params: { dragItem: DragItem; dropResult: DropResult }): void;

	/**
	 * Whether the target can be dropped or not.
	 */
	canDrop?(params: { dragItem: DragItem; hoveredItem: HoveredItem }): boolean;

	/**
	 * Handler when the dragging begins.
	 */
	onBeginDrag?(params: { dragItem: DragItem }): void;

	/**
	 * Handler when the dragging ends.
	 */
	onEndDrag?(params: { dragItem: DragItem; dropResult: DropResult | null }): void;

	/**
	 * Defines the kinds of dragItems this dropTarget accepts.
	 */
	acceptType?: string;
}

/**
 * If specified, the function will be called for each row and the returned value overrides key for the particular row.
 */
export type RowKeyGetter<RowType> = (params: { row: RowType }) => Key;

/**
 * Collections of callback function available on the row.
 */
export type RowEventHandlers = {
	/**
	 * A callback that will be triggered when the row is clicked.
	 */
	onClick?(event?: MouseEvent<HTMLElement>): void;
};

/**
 * If specified, it will be called for each row, and the returned functions will be called accordingly when event is
 * triggered on the row.
 */
export type RowEventHandlerGetter<RowType> = (params: { row: RowType; rowIndex: number }) => RowEventHandlers;

/**
 * Specify sort options: which column is active, sort order, and callback function for handling of sort event
 */
export interface SortOptions<ColumnType> {
	/**
	 * Values relevant to the current sorting state including: the column being sorted and the order/direction its sorted in.
	 */
	sortState?: SortState<ColumnType>;

	/**
	 * Handler function for sort event.
	 */
	onSort?(params: { column: ColumnType; order: SortOrder }): void;
}

/**
 * This interface is to couple a column with its sort order
 */
export interface SortState<ColumnType> {
	/**
	 * The column being sorted.
	 */
	column?: ColumnType;

	/**
	 * The direction the column is being sorted in.
	 */
	order?: SortOrder;
}

/**
 * This interface is to override default react-virtualized props/behaviors
 */
export interface VirtualScrollOptions extends Partial<ReactVirtualizedProps> {
	/**
	 * The reference of the react-virtualized list instance.
	 */
	listRef?: RefCallback<ReactVirtualizedList>;
}
// ...
export interface RowStyles extends Styleable {
	/**
	 * Title attribute for the row.
	 */
	title?: string;

	/**
	 * Whether the row is selected.
	 */
	selected?: boolean;

	/**
	 * Whether the row is interactive.
	 */
	interactive?: boolean;

	/**
	 * Whether the row is disabled.
	 */
	disabled?: boolean;

	/**
	 * Highlight row bases on different variables.
	 */
	highlightVariant?: TableTemplateProps.TableHighlightVariant;

	/**
	 * Whether the row is highlighted.
	 */
	highlighted?: boolean;

	/**
	 * Whether the context menu is disabled from some specific Rows.
	 */
	disabledRightClickContextMenu?: boolean;
}

/**
 * If specified, this function will be called for each row and return row style overriding default style.
 */
export type RowStyleGetter<RowType> = (params: { row: RowType; rowIndex: number }) => RowStyles;

/**
 * Same as {@link RowStyles}, but for cell.
 */
export interface CellStyles extends Styleable {
	/**
	 * Whether the body cell is secondary info.
	 */
	useSecondaryColor?: boolean;

	/**
	 * The hidden text will be placed at the secondary cell.
	 *
	 * @default placed at the first secondary cell of the row with content:
	 * - English: "Withdrawn"
	 * - German: "Löschen (wiederherstellbar)"
	 */
	secondaryCellTitle?: string;
}

/**
 * If specified, this function will be called for each cell and return cell style overriding default style.
 */
export type CellStyleGetter<RowType, ColumnType extends BaseColumnType<RowType>> = (params: {
	row: RowType;
	rowIndex: number;
	column: ColumnType;
}) => CellStyles;

/**
 * Hold common static prop values that are shared through all table components.
 */
export interface TableContextType<
	RowType = unknown,
	ColumnType extends BaseColumnType<RowType> = BaseColumnType<RowType>
> {
	rowKey?: string | number | RowKeyGetter<RowType>;
	columns: ColumnType[];
	flattenColumns?: ColumnType[];
	sortOptions?: SortOptions<ColumnType>;
	disabled?: boolean;
	cardView?: boolean;
	hasFootContent?: boolean;
	dragDropOptions?: TableDragDropOptions<RowType>;
	virtualScrollOptions?: true | Partial<ReactVirtualizedProps>;
	infiniteScrollOptions?: InfiniteScrollOptions;
	columnResizingOptions?: ColumnResizingOptions<ColumnType>;
	rowStyling?: RowStyleGetter<RowType>;
	cellStyling?: CellStyleGetter<RowType, ColumnType>;
	rowEventHandlers?: RowEventHandlerGetter<RowType>;
	componentRenderers: TableComponentRenderers<RowType, ColumnType>;
	resizable?: boolean;
	hasColumnGroup?: boolean;
	crossTabulation?: boolean;
	cellHighlighting?: boolean;
	isInteractiveTable?: boolean;
	setIsInteractiveTable?: Dispatch<SetStateAction<boolean>>;
	isDragging?: boolean;
	setIsDragging?: Dispatch<SetStateAction<boolean>>;
	hasScrollToNode?: boolean;
}

/** @deprecated since version 38.2.0. Use {@link ColumnResizeEventHandler} instead. */
export type ResizeEventHandler<ColumnType> = ColumnResizeEventHandler<ColumnType>;
export type ColumnResizeEventHandler<ColumnType> = (params: {
	resizedColumn: ColumnType;
	resizedWidthsGetter?: (column: ColumnType) => Column.Width | undefined;
	event: DraggableEvent;
	data: DraggableData;
}) => void;

export interface ColumnResizingOptions<ColumnType> {
	/**
	 * Handle event when the user starts to resize a column.
	 */
	onBeginResize?: ColumnResizeEventHandler<ColumnType>;

	/**
	 * Handle event when a column is being resized.
	 */
	onResize?: ColumnResizeEventHandler<ColumnType>;

	/**
	 * Handle event when the user has finished resizing a column.
	 */
	onEndResize?: ResizeEventHandler<ColumnType>;
}